feat : StationName VO 기능 구현
