+++ src/main/java/wooteco/subway/station/Station.java
@@ -1,41 +1,22 @@
 package wooteco.subway.station;
 
 import java.util.Objects;
-import java.util.regex.Pattern;
 
 public class Station {
-    private static final int MAX_NAME_LENGTH = 20;
-    private static final Pattern VALID_NAME_PATTERN = Pattern.compile("^[·가-힣a-zA-Z0-9\\(\\)\\s]*$");
-
     private Long id;
-    private String name;
-
-    public Station() {
-        this(null, null);
-    }
+    private StationName name;
 
     public Station(String name) {
         this(null, name);
     }
 
     public Station(Long id, final String name) {
-        String trimAndRemoveDuplicateBlankName = name.trim().replaceAll(" +", " ");
-        validateNameLength(trimAndRemoveDuplicateBlankName);
-        validateInvalidName(trimAndRemoveDuplicateBlankName);
+        this(id, new StationName(name));
+    }
+
+    public Station(Long id, StationName name) {
         this.id = id;
-        this.name = trimAndRemoveDuplicateBlankName;
-    }
-
-    private void validateNameLength(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException(String.format("역 이름은 %d자를 초과할 수 없습니다. 이름의 길이 : %d", MAX_NAME_LENGTH, name.length()));
-        }
-    }
-
-    private void validateInvalidName(String name) {
-        if (!VALID_NAME_PATTERN.matcher(name).matches()) {
-            throw new IllegalArgumentException(String.format("역 이름에 유효하지 않은 문자가 있습니다. 역 이름 : %s", name));
-        }
+        this.name = name;
     }
 
     public Long getId() {
@@ -43,7 +24,7 @@ public class Station {
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     @Override
